refactor and revert latest bug
