+++ pollinator/constants.py
@@ -22,5 +22,5 @@ model_index = (
 def available_models():
     return list(requests.get(model_index).json().values()) + [
         "no-gpu-test-image",
-        # "avatarclip",
+        "avatarclip",
     ]  # TODO .keys()
