+++ pollinator/main.py
@@ -5,7 +5,7 @@ import traceback
 import click
 from realtime.connection import Socket
 
-from pollinator import constants, process_msg
+from pollinator import constants, cog_handler
 from pollinator.constants import supabase, supabase_api_key, supabase_id
 from pollinator.process_msg import process_message
 
@@ -29,13 +29,13 @@ def finish_all_tasks():
 
 def get_task_from_db():
     """Scan the db for tasks that are not in progress. If there are none, return None
-    If there are many, return the olderst one for the currently process_msg.loaded_model.
-    If there are none for the process_msg.loaded_model, return the oldest."""
+    If there are many, return the olderst one for the currently cog_handler.loaded_model.
+    If there are none for the cog_handler.loaded_model, return the oldest."""
     data = (
         supabase.table(constants.db_name)
         .select("*")
         .eq("processing_started", False)
-        .eq("image", process_msg.loaded_model)
+        .eq("image", cog_handler.loaded_model)
         .order("request_submit_time")
         .execute()
     )
@@ -57,16 +57,16 @@ def get_task_from_db():
 
 def maybe_process(message):
     if (
-        message["image"] != process_msg.loaded_model
-        and process_msg.loaded_model is not None
+        message["image"] != cog_handler.loaded_model
+        and cog_handler.loaded_model is not None
     ):
         logging.info(
             "Message is not for this model, waiting a bit to give other workers a chance"
         )
         time.sleep(1)
     elif (
-        message["image"] != process_msg.loaded_model
-        and process_msg.loaded_model is None
+        message["image"] != cog_handler.loaded_model
+        and cog_handler.loaded_model is None
     ):
         logging.info("No model loaded, wait 0.5s to give other workers a chance")
         time.sleep(0.5)
