+++ pollinator/process_msg.py
@@ -1,142 +1,28 @@
-import base64
 import datetime as dt
 import json
 import logging
 import os
-import shutil
-import subprocess
-import time
 import traceback
-from mimetypes import guess_extension
-
-import psutil
-import requests
-from retry import retry
 
 from pollinator import constants
-from pollinator.constants import available_models, supabase, test_image
-from pollinator.ipfs_to_json import ipfs_subfolder_to_json
+from pollinator.constants import available_models, supabase
+from pollinator.ipfs_to_json import (
+    fetch_inputs,
+    BackgroundCommand,
+    prepare_output_folder,
+    clean_folder,
+    write_folder,
+)
+from pollinator.cog_handler import (
+    RunningCogModel,
+    kill_cog_model,
+    send_to_cog_container,
+)
+
 
 ipfs_root = os.path.abspath("/tmp/ipfs/")
 output_path = os.path.join(ipfs_root, "output")
 input_path = os.path.join(ipfs_root, "input")
-
-
-class BackgroundCommand:
-    def __init__(self, cmd):
-        self.cmd = cmd
-
-    def __enter__(self):
-        self.proc = subprocess.Popen(
-            f"exec {self.cmd}",
-            shell=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        return self.proc
-
-    def __exit__(self, type, value, traceback):
-        logging.info(f"Killing background command: {self.cmd}")
-        tree_kill(self.proc.pid)
-        try:
-            logs, errors = self.proc.communicate(timeout=2)
-            logs, errors = logs.decode("utf-8"), errors.decode("utf-8")
-            logging.info(f"   Logs: {logs}")
-            logging.error(f"   errors: {errors}")
-        except subprocess.TimeoutExpired:
-            pass
-
-
-loaded_model = None
-
-
-@retry(tries=450, delay=2)
-def cogmodel_can_start_healthy():
-    """Wait for the cogmodel to load and return a healthy status code
-    If no docker command is running anymore, throw an exception"""
-    logging.info("Checking if cogmodel is healthy...")
-
-    # check that cogmodel is a running as a containere
-    if "cogmodel" not in os.popen("docker ps").read():
-        logging.error("No running cogmodel found in docker ps. Exiting")
-        return False
-    # check that it is healthy. This step might fail and and be retried
-    response = requests.get("http://localhost:5000/")
-
-    print(os.popen("cat /tmp/ipfs/output/logs").read())
-    return response.status_code == 200
-
-
-@retry(tries=60, delay=1)
-def wait_for_docker_container(cog_cmd):
-    logging.info(cog_cmd)
-    os.system(cog_cmd)
-    # logging.info(os.popen(cog_cmd).read())
-    logging.error(f"Trying to find cog container: {os.popen('docker ps').read()}")
-    assert "cogmodel" in os.popen("docker ps").read()
-    with open("/tmp/ipfs/output/log", "r") as f:
-        if "is already in use by container" in f.read():
-            kill_cog_model()
-            raise Exception("container name cogmodel is already in use")
-
-
-class UnhealthyModel(Exception):
-    pass
-
-
-class RunningCogModel:
-    def __init__(self, image, output_path):
-        self.image = image
-        gpus = "--gpus all" if image != test_image else ""
-        # Start cog container
-        self.cog_cmd = (
-            f'bash -c "docker run --rm --name cogmodel -p 5000:5000 '
-            f"--mount type=bind,source={output_path},target=/outputs "
-            f'{gpus} {image} &> {output_path}/log" &'
-        )
-        logging.info(f"Initializing cog command: {self.cog_cmd}")
-
-    def __enter__(self):
-        global loaded_model
-        if loaded_model == self.image:
-            try:
-                assert (
-                    requests.get(
-                        "http://localhost:5000/",
-                    ).status_code
-                    == 200
-                )
-                logging.info(f"Model already loaded: {self.image}")
-                return
-            except:  # noqa
-                logging.info(f"Loaded model unhealthy, restarting: {self.image}")
-        kill_cog_model()
-        logging.info(f"Starting {self.image}: {self.cog_cmd}")
-        wait_for_docker_container(self.cog_cmd)
-        if not cogmodel_can_start_healthy():
-            raise UnhealthyModel()
-
-        loaded_model = self.image
-
-    def __exit__(self, type, value, traceback):
-        pass
-
-
-@retry(tries=60, delay=1)
-def wait_until_cogmodel_is_free():
-    logging.info("docker kill cogmodel")
-    os.system("docker kill cogmodel")
-    assert "cogmodel" not in os.popen("docker ps").read()
-    logging.info("cogmodel killed and is container name is free.")
-    time.sleep(3)
-
-
-def kill_cog_model():
-    try:
-        wait_until_cogmodel_is_free()
-        logging.info(f"Killed previous model ({loaded_model})")
-    except Exception as e:  # noqa
-        logging.error(f"Error killing cogmodel: {type(e)}{e}")
 
 
 def process_message(message):
@@ -156,22 +42,18 @@ def process_message(message):
 
     try:
         updated_message["end_time"] = dt.datetime.now().strftime("%Y-%m-%dT%H:%M:%S")
-        data = (
-            supabase.table(constants.db_name)
-            .select("*")
-            .eq("input", message["input"])
-            .execute()
-            .data
-        )
-        assert len(data) == 1
-        cid = data[0]["output"]
-
+        
         data = (
             supabase.table(constants.db_name)
             .update(updated_message)
             .eq("input", message["input"])
             .execute()
+            .data
         )
+
+        assert len(data) == 1
+        cid = data[0]["output"]
+        # todo get cid from data
         print("Pollen set to done in db: ", data)
         logging.info(f"Got CID: {cid}. Triggering pinning and social post")
         # run pinning and social post
@@ -205,7 +87,6 @@ def start_container_and_perform_request_and_send_outputs(message):
     clean_folder(input_path)
     prepare_output_folder(output_path)
     inputs = fetch_inputs(message["input"])
-
     # Write inputs to /input
     for key, value in inputs.items():
         write_folder(input_path, key, json.dumps(value))
@@ -228,102 +109,3 @@ def start_container_and_perform_request_and_send_outputs(message):
         f"| python pollinator/outputs_to_db.py {message['input']} {constants.db_name}",
     )
     return message, success
-
-
-def prepare_output_folder(output_path):
-    logging.info(f"Mounting output folder: {output_path}")
-    os.makedirs(output_path, exist_ok=True)
-    clean_folder(output_path)
-    write_folder(output_path, "done", "false")
-    write_folder(output_path, "time_start", str(int(time.time())))
-
-
-def fetch_inputs(ipfs_cid):
-    try:
-        inputs = ipfs_subfolder_to_json(ipfs_cid, "input")
-    except KeyError:
-        raise ValueError(f"IPFS hash {ipfs_cid} could ot be resolved")
-    logging.info(f"Fetched inputs from IPFS {ipfs_cid}: {inputs}")
-    return inputs
-
-
-def send_to_cog_container(inputs, output_path):
-    logging.info("Send to cog model")
-    # Send message to cog container
-    payload = {"input": inputs}
-    response = requests.post("http://localhost:5000/predictions", json=payload)
-
-    logging.info(f"response: {response}")
-
-    write_folder(output_path, "time_start", str(int(time.time())))
-
-    if response.status_code != 200:
-        logging.error(response.text)
-        write_folder(output_path, "cog_response", response.text, "a")
-        kill_cog_model()
-        raise Exception(
-            f"Error while sending message to cog container: {response.text}"
-        )
-    else:
-        write_http_response_files(response, output_path)
-        write_folder(output_path, "done", "true")
-        logging.info(f"Set done to true in {output_path}")
-
-    return response
-
-
-# Since ipfs reads its data from the filesystem we write keys and values to files using this function
-# TODO: needs to handle URL values
-def write_folder(path, key, value, mode="w"):
-    os.makedirs(path, exist_ok=True)
-    with open(f"{path}/{key}", mode) as f:
-        f.write(value)
-
-
-def clean_folder(folder):
-    os.makedirs(folder, exist_ok=True)
-    for filename in os.listdir(folder):
-        file_path = os.path.join(folder, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print("Failed to delete %s. Reason: %s" % (file_path, e))
-
-
-def write_http_response_files(response, output_path):
-    try:
-        output = response.json()["output"]
-        if not isinstance(output, list):
-            output = [output]
-        for i, encoded_file in enumerate(output):
-            try:
-                encoded_file = encoded_file["file"]
-            except TypeError:
-                pass  # already a string
-            meta, encoded = encoded_file.split(";base64,")
-            extension = guess_extension(meta.split(":")[1])
-            with open(f"{output_path}/out_{i}{extension}", "wb") as f:
-                f.write(base64.b64decode(encoded))
-    except Exception as e:  # noqa
-        logging.info(f"http response not written to file: {type(e)} {e}")
-        traceback.print_exc()
-
-
-def tree_kill(pid):
-    print(f"Killing process {pid} and their complete family")
-    parent = psutil.Process(pid)
-    for child in parent.children(recursive=True):
-        print(f"Killing child: {child} {child.pid}")
-        child.kill()
-    parent.kill()
-
-
-# if __name__ == "__main__":
-#     message = {
-#         "input": "QmNgrCgddkXpRhiZVYHuuuM5KCu4DJDPP9F1K9kW99etfY",
-#         "image": "majesty"
-#     }
-#     process_message(message)
